refactor: 함수 네이밍 및 경로 수정
