+++ src/controller/handleCountInput.js
@@ -1,8 +1,8 @@
 import { $ } from '../utils/selector.js';
 import { ERROR_MESSAGES } from '../utils/constants.js';
 import { isValidCount } from '../utils/valid.js';
+import { setDisabledAllForms } from '../view/setFormState.js';
 import handleRace from './handleRace.js';
-import { setDisabledAllForms } from '../view/showProgress.js';
 
 const handleCountInput = () => {
   const { value } = $('.count-input');
