+++ src/utils/wait.js
@@ -1,0 +1,3 @@
+const wait = ms => new Promise(resolve => setTimeout(resolve, ms));
+
+export default wait;
