--- src/utils/waitSeconds.js
@@ -1,3 +1,0 @@
-const waitSeconds = ms => new Promise(resolve => setTimeout(resolve, ms));
-
-export default waitSeconds;
